DFR-3080: Added the refactored applicant field code so that it applies to paper cases, originally introduced with DFR-3068
